requirements.txt
pygame
pygame_gui
fastapi
uvicorn
sqlmodel
requests
passlib[bcrypt]
python-multipart
python-jose[cryptography]
paho-mqtt
websockets